removed registry
