--- registry/alembic/versions/89b5ef5289ee_changed_member_table_dependency_to_org_.py
@@ -1,34 +1,0 @@
-"""changed member table dependency to org_uuid
-
-Revision ID: 89b5ef5289ee
-Revises: ff01af45872f
-Create Date: 2020-01-15 15:33:22.638996
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = '89b5ef5289ee'
-down_revision = 'ff01af45872f'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('org_member', sa.Column('created_on', mysql.TIMESTAMP(), nullable=True))
-    op.add_column('org_member', sa.Column('transaction_hash', sa.VARCHAR(length=128), nullable=True))
-    op.add_column('org_member', sa.Column('updated_on', mysql.TIMESTAMP(), nullable=True))
-    op.drop_column('org_member', 'org_row_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('org_member', sa.Column('org_row_id', mysql.INTEGER(display_width=11), autoincrement=False, nullable=False))
-    op.drop_column('org_member', 'updated_on')
-    op.drop_column('org_member', 'transaction_hash')
-    op.drop_column('org_member', 'created_on')
-    # ### end Alembic commands ###
