--- registry/alembic/versions/9ab2f425904b_added_unique_constraint_on_the_member_.py
@@ -1,30 +1,0 @@
-"""added unique constraint on the member username & org_uuid
-
-Revision ID: 9ab2f425904b
-Revises: 89b5ef5289ee
-Create Date: 2020-01-21 13:06:43.309986
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = '9ab2f425904b'
-down_revision = '89b5ef5289ee'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('org_member', sa.Column('invited_on', mysql.TIMESTAMP(), nullable=True))
-    op.create_unique_constraint('uk_user_org', 'org_member', ['org_uuid', 'username'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('uk_user_org', 'org_member', type_='unique')
-    op.drop_column('org_member', 'invited_on')
-    # ### end Alembic commands ###
