--- registry/alembic/versions/ff01af45872f_add_org_member_table.py
@@ -1,35 +1,0 @@
-"""add org member table
-
-Revision ID: ff01af45872f
-Revises: de9162a33a08
-Create Date: 2020-01-13 14:01:42.257841
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'ff01af45872f'
-down_revision = 'de9162a33a08'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('org_member', sa.Column('address', sa.VARCHAR(length=128), nullable=True))
-    op.add_column('org_member', sa.Column('invite_code', sa.VARCHAR(length=128), nullable=True))
-    op.add_column('org_member', sa.Column('org_row_id', sa.Integer(), nullable=False))
-    op.add_column('org_member', sa.Column('status', sa.VARCHAR(length=128), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'org_member', type_='foreignkey')
-    op.drop_column('org_member', 'status')
-    op.drop_column('org_member', 'org_row_id')
-    op.drop_column('org_member', 'invite_code')
-    op.drop_column('org_member', 'address')
-    # ### end Alembic commands ###
