--- registry/config.py
@@ -1,37 +1,0 @@
-NETWORKS = {
-    3: {
-        "name": "test",
-        "http_provider": "https://ropsten.infura.io",
-        "ws_provider": "wss://ropsten.infura.io/ws",
-        "db": {
-            "DB_HOST": "localhost",
-            "DB_USER": "unittest_root",
-            "DB_PASSWORD": "unittest_pwd",
-            "DB_NAME": "registry_unittest_db",
-            "DB_PORT": 3306,
-        },
-    }
-}
-NETWORK_ID = 3
-SLACK_HOOK = {
-    'hostname': '',
-    'port': 443,
-    'path': '',
-    'method': 'POST',
-    'headers': {
-        'Content-Type': 'application/json'
-    }
-}
-IPFS_URL = {
-    'url': '',
-    'port': '80',
-
-}
-ALLOWED_ORIGIN = ["PUBLISHER_PORTAL"]
-METADATA_FILE_PATH = "/tmp"
-REGION_NAME = "us-east-1"
-ASSET_BUCKET = ""
-ASSET_DIR = "/tmp"
-NOTIFICATION_ARN = ""
-PUBLISHER_PORTAL_DAPP_URL = ""
-
