--- registry/consumer/event_consumer.py
@@ -1,4 +1,0 @@
-
-class EventConsumer(object):
-    def on_event(self, event):
-        pass
