--- registry/domain/models/service.py
@@ -1,3 +1,0 @@
-
-class Service:
-    pass
