--- registry/requirements.txt
@@ -1,8 +1,0 @@
-alembic==1.3.1
-boto3==1.10.9
-ipfsapi==0.4.3
-PyMySQL==0.9.3
-SQLAlchemy==1.3.11
-web3==4.9.2
-deepdiff==4.0.9
-requests==2.22.0
