--- registry/testcases/unit_testcases/test_org_factory.py
@@ -1,5 +1,0 @@
-import unittest
-
-
-class TestOrgFactory(unittest.TestCase):
-    pass
